python: Optimize completion for loading lazily on demand
Move `pip` to separate completion definition delegating to pip to do
all the completion work.

Since the completion is loaded lazily on demand, it avoids the
performance overhead during Zsh initialization.

Implementation note:
The helper funtion `_pip_completion` implementation is based on the
the official pip completion function (which can be generated with
`pip completion --zsh`) to adhere to the newer compsys style.

See: 'man zshcompsys' for more details.